requirements.txt
# generated from manifests external_dependencies
odoorpc
OERPLib-py3
